Penambahan file testing dan factory

+++ app/Http/Controllers/VideoController.php
@@ -111,14 +111,20 @@ class VideoController extends Controller
 
     public function destroy($id): RedirectResponse
     {
-        $video = Videos::findOrFail($id);
+        $video = Videos::find($id);
 
-        // Hapus file video jika ada
+        if (!$video) {
+            return redirect()->route('videos.index')->with(['error' => 'Video not found.']);
+        }
+
+        if (Auth::id() !== $video->uploader_id) {
+            return redirect()->route('videos.index')->with(['error' => 'Tidak bisa menghapus video milik orang lain']);
+        }
+
         if ($video->video && Storage::exists('public/videos/' . $video->video)) {
             Storage::delete('public/videos/' . $video->video);
         }
 
-        // Hapus entri video dari database
         $video->delete();
 
         return redirect()->route('videos.index')->with(['success' => 'Video Berhasil Dihapus!']);
@@ -127,7 +133,12 @@ class VideoController extends Controller
     public function incrementViews($id)
     {
         $video = Videos::findOrFail($id);
+
+        if ($video->uploader_id == Auth::id()) {
+            return response()->json(['success' => false, 'message' => 'Pemilik video tidak dapat menambah jumlah tayangan.'], 403);
+        }
         $video->increment('views');
+
         return response()->json(['success' => true]);
     }
 }
